+++ errors.go
@@ -1,3 +1,7 @@
+// Copyright 2022 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
 package fincen
 
 import "fmt"
