+++ suspicious_activity/primitive.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by an Apache License
 // license that can be found in the LICENSE file.
 
-package form8300
+package suspicious_activity
 
 import (
 	"reflect"
@@ -10,12 +10,12 @@ import (
 	"github.com/moov-io/fincen"
 )
 
-// May be one of 1
+// May be one of 1, 2, 3, 4, 5
 type ValidateActivityNarrativeSequenceNumber int
 
 func (r ValidateActivityNarrativeSequenceNumber) Validate() error {
 	for _, vv := range []string{
-		"1",
+		"1", "2", "3", "4", "5",
 	} {
 		if reflect.DeepEqual(string(r), vv) {
 			return nil
@@ -24,12 +24,12 @@ func (r ValidateActivityNarrativeSequenceNumber) Validate() error {
 	return fincen.NewErrValueInvalid("ValidateActivityNarrativeSequenceNumber")
 }
 
-// May be one of 35, 37, 16, 23, 4, 8, 3
+// May be one of 35, 37, 30, 33, 34, 8, 46, 18, 19, 41
 type ValidateActivityPartyCodeType string
 
 func (r ValidateActivityPartyCodeType) Validate() error {
 	for _, vv := range []string{
-		"35", "37", "16", "23", "4", "8", "3",
+		"35", "37", "30", "33", "34", "8", "46", "18", "19", "41",
 	} {
 		if reflect.DeepEqual(string(r), vv) {
 			return nil
@@ -38,12 +38,12 @@ func (r ValidateActivityPartyCodeType) Validate() error {
 	return fincen.NewErrValueInvalid("ValidateActivityPartyCodeType")
 }
 
-// May be one of 1, 2, 5, 6, 7, 28, 999
+// May be one of 1, 2, 4, 5, 6, 7, 9, 10, 11, 12, 13, 14, 28, 32, 33, 29, 999
 type ValidatePartyIdentificationCodeType string
 
 func (r ValidatePartyIdentificationCodeType) Validate() error {
 	for _, vv := range []string{
-		"1", "2", "5", "6", "7", "28", "999",
+		"1", "2", "4", "5", "6", "7", "9", "10", "11", "12", "13", "14", "28", "32", "33", "29", "999",
 	} {
 		if reflect.DeepEqual(string(r), vv) {
 			return nil
@@ -52,16 +52,30 @@ func (r ValidatePartyIdentificationCodeType) Validate() error {
 	return fincen.NewErrValueInvalid("ValidatePartyIdentificationCodeType")
 }
 
-// May be one of L, DBA
+// May be one of L, AKA, DBA
 type ValidatePartyNameCodeType string
 
 func (r ValidatePartyNameCodeType) Validate() error {
 	for _, vv := range []string{
-		"L", "DBA",
+		"L", "AKA", "DBA",
 	} {
 		if reflect.DeepEqual(string(r), vv) {
 			return nil
 		}
 	}
 	return fincen.NewErrValueInvalid("ValidatePartyNameCodeType")
+}
+
+// May be one of 9, 1, 2, 7, 3, 4, 6, 13, 99
+type ValidateFederalRegulatorCodeType string
+
+func (r ValidateFederalRegulatorCodeType) Validate() error {
+	for _, vv := range []string{
+		"9", "1", "2", "7", "3", "4", "6", "13", "99",
+	} {
+		if reflect.DeepEqual(string(r), vv) {
+			return nil
+		}
+	}
+	return fincen.NewErrValueInvalid("ValidateFederalRegulatorCodeType")
 }
